requirements.txt
annotated-types==0.7.0
anthropic==0.39.0
anyio==4.12.1
APScheduler==3.10.4
bcrypt==4.2.1
cachetools==5.5.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.5
defusedxml==0.7.1
distro==1.9.0
et_xmlfile==2.0.0
fonttools==4.62.1
fpdf2==2.8.2
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.157.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.1
google-genai==1.68.0
google-generativeai==0.8.6
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.71.2
gspread==6.2.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
jiter==0.13.0
lxml==6.0.2
oauthlib==3.3.1
openpyxl==3.1.5
pillow==12.1.1
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-calamine==0.6.2
python-docx==1.2.0
python-dotenv==1.0.1
python-telegram-bot==21.5
pytz==2024.1
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
tenacity==9.1.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
websockets==16.0
xlrd==2.0.2
